Invoke repair_entities() in more activity modules, refactor
